Update from Master

+++ prepareData/DownloadData.py
@@ -31,29 +31,59 @@
 
 from urllib.request import urlretrieve
 import zipfile 
+import os.path
+import sys
 
 def downloadAndExtractData():
 
-    print('Download data files')
+    print('Download data files', flush=True)
 
-    print('    Download hla_nom_p.txt')
-    urlretrieve('http://hla.alleles.org/wmda/hla_nom_p.txt', filename='hla_nom_p.txt')
+    successStatus = 0
 
-    print('    Download hla_nom_g.txt')
-    urlretrieve('http://hla.alleles.org/wmda/hla_nom_g.txt', filename='hla_nom_g.txt')
+    try:
+        if os.path.isfile('hla_nom_p.txt'):
+            print('    File hla_nom_p.txt already present in directory', flush=True)
+        else:
+            print('    Download hla_nom_p.txt', flush=True)
+            urlretrieve('http://hla.alleles.org/wmda/hla_nom_p.txt', filename='hla_nom_p.txt')
+    
+        if os.path.isfile('hla_nom_g.txt'):
+            print('    File hla_nom_g.txt already present in directory', flush=True)
+        else:
+            print('    Download hla_nom_g.txt', flush=True)
+            urlretrieve('http://hla.alleles.org/wmda/hla_nom_g.txt', filename='hla_nom_g.txt')
 
-    print('    Download alpha.v3.zip')
-    urlretrieve('https://bioinformatics.bethematchclinical.org/HLA/alpha.v3.zip', filename='alpha.v3.zip')
-    print('    Extract alpha.v3.zip')
-    alphav3Zip = zipfile.ZipFile('alpha.v3.zip')
-    alphav3Zip.extract('alpha.v3.txt')
+        if os.path.isfile('alpha.v3.zip'):
+            print('    File alpha.v3.zip is already present in directory', flush=True)
+        else:
+            print('    Download alpha.v3.zip', flush=True)
+            urlretrieve('https://bioinformatics.bethematchclinical.org/HLA/alpha.v3.zip', filename='alpha.v3.zip')
 
-    print('    Download hla_ambigs_xml.zip')
-    urlretrieve('https://github.com/jrob119/IMGTHLA/raw/Latest/xml/hla_ambigs.xml.zip', filename='hla_ambigs.xml.zip')
-    print('    Extract hla_ambigs_xml.zip')
-    hlaAmbigZip = zipfile.ZipFile('hla_ambigs.xml.zip')
-    hlaAmbigZip.extract('hla_ambigs.xml')
+        if os.path.isfile('hla_ambigs.xml.zip'):
+            print('    File hla_ambigs.xml.zip is already present in directory', flush=True)
+        else:
+            print('    Download hla_ambigs.xml.zip', flush=True)
+            urlretrieve('https://github.com/jrob119/IMGTHLA/raw/Latest/xml/hla_ambigs.xml.zip', filename='hla_ambigs.xml.zip')
 
+    except:
+        sys.stderr.write('Something went wrong downloading files. You might want to check proxies or firewalls. You can also manually download the files.')
+        successStatus = 1
+
+    if successStatus < 1:
+        try:
+            print('    Extract alpha.v3.zip', flush=True)
+            alphav3Zip = zipfile.ZipFile('alpha.v3.zip')
+            alphav3Zip.extract('alpha.v3.txt')
+
+            print('    Extract hla_ambigs_xml.zip', flush=True)
+            hlaAmbigZip = zipfile.ZipFile('hla_ambigs.xml.zip')
+            hlaAmbigZip.extract('hla_ambigs.xml')
+
+        except:
+            sys.stderr.write('Something went wrong extracting files.')
+            successStatus = 2
+
+    return successStatus
 
 if __name__ == "__main__":
